Se reorganiza la estructura de archivos para seguir una de las convenciones más usuales
